create small-ish training set
I now have a dataframe of a small selection of poets' work to test the fine-tuning openai function with. In order to use it, I will have to trim it down to just poems and save it as a csv file then use openai's tool to format the file. Then all that remains is to fine-tune.